将DUGame、glog导出到lua

+++ server/server/main.cpp
@@ -1,55 +1,40 @@
-
-#include <iostream>
-#include <chrono>
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-#include "lua.h"
-#include "lualib.h"
-#include "lauxlib.h"
-#ifdef __cplusplus
-}
-#endif
-
-
 #include "lua-modules/lua_modules.h"
-
 #include "DUGame.h"
 
-
-void main()
+void lua_pre_register_all(lua_State* L)
 {
-	lua_State* L = luaL_newstate();
-
 	luaL_openlibs(L);
 	preload_lua_modules(L);
 
-	luaL_dofile(L, "lua_code/main.lua");
+	const luaL_reg global_functions[] =
+	{
+		{ "print", lua_print },
+		{ "release_print",lua_release_print },
+		{ nullptr, nullptr }
+	};
+	luaL_register(L, "_G", global_functions);
+}
+
+void main()
+{
+	const char* startlua = "lua_code/main.lua";
 
 	auto instance = DUGame::getInstance();
+
+	if (!instance->isFileExist(startlua))
+	{
+		return;
+	}
+
+	lua_State* L = instance->luaState();
+	
+	lua_pre_register_all(L);
+
+	luaL_dofile(L, startlua);
+
 	instance->setInterval(1 / 40.0f);
 	instance->run();
 
 	DUGame::destroy();
-
-	lua_close(L);
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
